backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.14"
authors = [{ name = "Konstantin Kiselev", email = "konstantin.kiselev.1998@mail.ru" }]
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.3",
    "gunicorn>=25.1.0",
    "pillow>=12.1.1",
    "psycopg>=3.3.3",
    "python-dotenv>=1.2.1",
    "pytils>=0.4.4",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "django-extensions>=4.1",
    "django-stubs>=5.2.9",
    "ipython>=9.10.0",
    "mypy>=1.19.1",
    "mypy-extensions>=1.1.0",
    "pre-commit>=4.5.1",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 79
target-version = "py314"
extend-exclude = [
    ".env",
    ".git",
    ".idea",
    "venv",
    "**/migrations/*.py",
    "**/settings*.py",
    "*/settings*.py",
]

[tool.ruff.format]
indent-style = "space"
quote-style = "single"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "TD002", # missing author in TODO
    "TD003", # missing issue link in TODO
    "TD004", # missing colon in TODO
    "FIX002", # line contains TODO/FIXME/XXX

    "ARG001", #, Unused function args

    "Q000", # double quotes found but single quotes preferred
    "Q003", # change outer quotes to avoid escaping inner quotes

    "ANN", # let mypy handle it

    "E501", # line too long

    "COM812", # trailing comma missing
    "COM819", # trailing comma prohibited

    "B904", # use raise from to specify exception cause

    "D100", # missing docstring in public module
    "D101", # missing docstring in public class
    "D102", # missing docstring in public method
    "D103", # missing docstring in public function
    "D104", # missing docstring in public package
    "D105", # missing docstring in magic method
    "D106", # missing docstring in public nested class
    "D107", # missing docstring in `__init__`
    "D200", # one-line docstring should fit on one line
    "D202", # no blank lines allowed after function docstring
    "D203", # 1 blank line required before class docstring
    "D205", # 1 blank line required between summary line and description
    "D209", # multi-line docstring closing quotes on separate line
    "D210", # no whitespaces around docstring text
    "D212", # multi-line docstring summary should start at first line
    "D213", # multi-line docstring summary should start at second line
    "D400", # first line should end with a period
    "D401", # first line should be in imperative mood
    "D404", # first word should not be "This"
    "D415", # first line should end with punctuation

    "EM101", # exception must not use a string literal
    "EM102", # exception must not use an f-string literal

    "INT001", # i18n: f-string resolved before call

    "N806", # variable in function should be lowercase

    "PGH003", # use specific rule codes when ignoring type issues
    "PGH004", # use specific rule codes when using noqa

    "RUF001", # ambiguous unicode in string
    "RUF002", # ambiguous unicode in docstring
    "RUF005", # consider iterable unpacking instead of concatenation
    "RUF012", # mutable class attributes should be ClassVar
    "RUF003", # ambiguous unicode character in comment
    "RUF100", # Unused noqa directive

    "PLR0913", # too many arguments

    "PTH118", # os.path.join -> Path / operator
    "PTH119", # os.path.basename -> Path.name
    "PTH120", # os.path.dirname -> Path.parent
    "PTH122", # os.path.splitext -> Path.suffix
    "PTH123", # open() -> Path.open()

    "RET501", # do not explicitly return None if only possible return value
    "RET502", # do not implicitly return None if function can return non-None
    "RET503", # missing explicit return at end of function

    "RUF015", # prefer next(...) over [0] slice

    "S101", # use of assert

    "SIM108", # ternary operator instead of if-else

    "TC001", # move app import into TYPE_CHECKING
    "TC002", # move third-party import into TYPE_CHECKING
    "TC003", # move stdlib import into TYPE_CHECKIN

    "SLF001", # private member accessed

    "TRY003", # avoid long messages outside exception class

    "UP015", # unnecessary open mode params
    "UP017", # use datetime.UTC
]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "pydantic",
    "ninja",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]
known-local-folder = ["apps", "api", "project"]
combine-as-imports = true
force-wrap-aliases = true
split-on-trailing-comma = true
lines-after-imports = 2

[tool.ruff.lint.isort.sections]
django = ["django"]
ninja = ["ninja"]
pydantic = ["pydantic"]

[tool.ruff.lint.per-file-ignores]
"*/factory.py" = ["ANN"]
"*/fixtures.py" = ["ANN"]
"*/fixtures/*" = ["ANN"]
"*/management/*" = ["ANN"]
"*/migrations/*" = ["ANN"]
"*/tests/*" = ["ANN"]

[tool.mypy]
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]
show_column_numbers = true
strict = true

[tool.django-stubs]
django_settings_module = "project.settings"
strict_model_abstract_attrs = true
strict_settings = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "project.settings",
    "apps.users.migrations",
    "apps.devices.migrations",
]
ignore_errors = true
